+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-10-13
+
+### Added
+- fwtools version to build configuration file (synthesis.py)
+- used utils.get_colored_logger (in synthesis.py and fwpacker.py)
+
 ## [0.2.0] - 2023-10-10
 
 ### Added
@@ -30,7 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - migrated scripts from `ugt_mp7_legacy/scripts` repo.
 
-[Unreleased]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.2.0...HEAD
+[Unreleased]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/cms-l1-globaltrigger/ugt-fwtools/compare/0.1.0...0.1.1